// src/lib/weebdex/model/getMangaIdChaptersParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * WeebDex API
 * WeebDex is an ad-free, scanlator-first online manga reader providing high-quality, zero-compression images. <br /><br />
Contact: [delta@weebdex.org](mailto:delta@weebdex.org) or join our [Discord Server](https://discord.gg/syp3a6Eucq).<br /><br />
All API endpoints have a global rate limit of 5 requests per second per IP. <br />
To avoid future issues, include the Origin: https://weebdex.org and Referer: https://weebdex.org/ headers when making API requests.<br/>
 * OpenAPI spec version: 1.2.0
 */
import type { GetMangaIdChaptersOrder } from "./getMangaIdChaptersOrder";
import type { GetMangaIdChaptersSort } from "./getMangaIdChaptersSort";

export type GetMangaIdChaptersParams = {
  /**
   * Groups to include (array of IDs)
   * @maxItems 25
   */
  group?: string[];
  /**
   * Groups to exclude (array of IDs)
   * @maxItems 25
   */
  groupx?: string[];
  /**
   * @minimum 0
   * @maximum 500
   */
  limit?: number;
  order?: GetMangaIdChaptersOrder;
  /**
   * @minimum 1
   */
  page?: number;
  sort?: GetMangaIdChaptersSort;
  /**
   * Translated language(s) to include
   */
  tlang?: string[];
  /**
   * Translated language(s) to exclude
   */
  tlangx?: string[];
  /**
   * Uploaders to include (array of IDs)
   * @maxItems 25
   */
  uploader?: string[];
  /**
   * Uploaders to exclude (array of IDs)
   * @maxItems 25
   */
  uploaderx?: string[];
};